feat : create money
